+++ src/App.jsx
@@ -5,7 +5,7 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import HomePage from './Pages/HomePage';
 import AboutPage from './Pages/AboutPage';
 import Contact from './Pages/ContactPage';
-import CarList from './Pages/CarList';
+import CarListPage from './Pages/CarListPage';
 import SingleCarPage from './Pages/SingleCarPage';
 import PrivacyPolicy from './Pages/PrivacyPolicy'
 
@@ -18,7 +18,7 @@ function App() {
         <Route path="/" element={<HomePage />} />
         <Route path="/about" element={<AboutPage />} />
         <Route path="/contact" element={<Contact />} />
-        <Route path="/carlist" element={<CarList />} />
+        <Route path="/carlist" element={<CarListPage />} />
         <Route path="/carlist/:numerodechassi" element={<SingleCarPage />} />
         <Route path="/privacypolicy" element={<PrivacyPolicy />} />
       </Routes>
