+++ src/Pages/AboutPage.jsx
@@ -9,7 +9,7 @@ function AboutPage() {
           About us
         </h1>
       </div>
-      <div className="grid p-10 flex container mx-auto grid-cols-2 gap-5">
+      <div className="grid p-10 flex container mx-auto grid-cols-2">
       <img className="rounded-lg" src="/aboutimg.jpg" alt="Banner" />
         <p className="p-5 text-lg w-4/5">
           Lorem ipsum dolor sit amet consectetur adipisicing elit. Dolores
